Fix capture dispatch overflow on large HP

+++ src/Aion2Flow.SceneRuntime/Archive/SceneArchivePayload.cs
@@ -787,13 +787,13 @@ public sealed class SceneArchiveEntityIdentity
     public NpcKind Kind { get; init; }
     public int? OwnerEntityId { get; init; }
     public EntityOwnerKind OwnerKind { get; init; }
-    public int? CurrentHp { get; init; }
-    public int? MaxHp { get; init; }
+    public long? CurrentHp { get; init; }
+    public long? MaxHp { get; init; }
     public bool NpcCombatActive { get; init; }
-    public uint? Value2136 { get; init; }
-    public uint? Sequence2136 { get; init; }
-    public uint? Value0140 { get; init; }
-    public uint? Value0240 { get; init; }
+    public long? Value2136 { get; init; }
+    public long? Sequence2136 { get; init; }
+    public long? Value0140 { get; init; }
+    public long? Value0240 { get; init; }
     public (byte State0, byte State1)? State4636 { get; init; }
     public (int SequenceId, int ResultCode)? Latest2C38 { get; init; }
     public long LastObservedOrdinal { get; init; }
@@ -840,8 +840,8 @@ public sealed class SceneArchiveEntityIdentity
 public sealed class SceneArchiveBossFocus
 {
     public int InstanceId { get; init; }
-    public int Hp { get; init; }
-    public int MaxHp { get; init; }
+    public long Hp { get; init; }
+    public long MaxHp { get; init; }
     public long CumulativeLostHp { get; init; }
     public long LastObservedAtMilliseconds { get; init; }
     public bool HasHp { get; init; }
